Funcitonality:
Changed the way the Cookie system acts. Now it should enter cookie credentials in the fields if set. And be able to log in on other accounts if the textbox value get changed. during the process. Changing the cookie if chosen to remember again.

Design:
Changed the layout on the login page to have the controls outlined better.